+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0 - xxxx-xx-xx
 - [CSE] Added support for 'm2m:dbg' throughout the CSE.
 - [CSE] Added configuration and command line argument to enable and disable validation of attributes and arguments.
+- [CSE] Added import of attribute policies of validation of &lt;flexContainer> specializations.
 - [TESTS] Added test cases for unit tests.